What a terrible mess.
